pyproject.toml
[tool.poetry]
name = "telegram-acr-lyrics-bot"
version = "0.1.0"
description = ""
authors = ["Andriy Orehov <andriyorehov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
pytelegrambotapi = "^3.6"
requests = "^2.22"
raven = "^6.10"
beautifulsoup4 = "^4.8"
google-measurement-protocol = "^1.0"
amalgama = "^0.2.2"
acr-cloud = "^0.1.2"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
